feat: split feed components into separate files
