+++ app/profile/[[...username]]/page.tsx
@@ -1,5 +1,4 @@
-import { getServerSession, User } from "next-auth";
-
+import { getFeedContent } from "lib/feed";
 import { getUsername, getUser } from "utils/getUser";
 
 import FeedPage from "../../../components/feed/FeedPage";
@@ -7,11 +6,6 @@ import PageTitle from "../../../components/ui/PageTitle";
 import ProfileImg from "../../../components/userProfile/profileImg";
 import { formatFollowers } from "../../../lib/formatters";
 import getSpotifyData from "../../../lib/getSpotifyData";
-import { authOptions } from "../../../pages/api/auth/[...nextauth]";
-import { getFeedContent } from "../../../pages/api/feedContent/[id]";
-import db from "../../firebase";
-
-import type { FeedItemContent } from "../../../components/feed/FeedPage";
 
 const getUserProfile = async (username: string) =>
   getSpotifyData<SpotifyApi.UserProfileResponse>(
