+++ lib/feed/index.ts
@@ -1,0 +1,6 @@
+import getFeedComments from "./getFeedComments";
+import getFeedContent from "./getFeedContent";
+import postFeedComment from "./postFeedComment";
+import postFeedContent from "./postFeedContent";
+
+export { getFeedComments, getFeedContent, postFeedComment, postFeedContent };
